Fix numpy float deprecated
On numpy version 1.24.0, numpy.float, numpy.int and similar aliases were deprecated.
Now we must to use float, int, etc. instead.
https://levelup.gitconnected.com/fix-attributeerror-module-numpy-has-no-attribute-float-d7d68c5a4971

+++ classes/pyETM.py
@@ -2667,7 +2667,7 @@ class ETM:
             f = np.ones((l.shape[1],))
 
             sw = np.power(10, LIMIT - ss[ss > LIMIT])
-            sw[sw < np.finfo(np.float).eps] = np.finfo(np.float).eps
+            sw[sw < np.finfo(float).eps] = np.finfo(float).eps
             f[ss > LIMIT] = sw
 
             p.append(np.square(np.divide(f, factor[i])))
@@ -2728,7 +2728,7 @@ class ETM:
                 # (missing jump, etc) you end up with very unstable inversions
                 # f[f > 500] = 500
                 sw = np.power(10, LIMIT - s[s > LIMIT])
-                sw[sw < np.finfo(np.float).eps] = np.finfo(np.float).eps
+                sw[sw < np.finfo(float).eps] = np.finfo(float).eps
                 f[s > LIMIT] = sw
 
                 P = np.square(np.divide(f, factor))
@@ -2736,7 +2736,7 @@ class ETM:
                 break  # cst_pass = True
 
         # make sure there are no values below eps. Otherwise matrix becomes singular
-        P[P < np.finfo(np.float).eps] = 1e-6
+        P[P < np.finfo(float).eps] = 1e-6
 
         # some statistics
         SS = np.linalg.inv(np.dot(A.transpose(), np.multiply(P[:, None], A)))
